Added browserify to gulp task
In order to be able to deploy to a browser it is necessary to compile to
a single bundle.js. Previously this was being done manually on the
command line after each transpile. This was getting rather tiresome so a
gulp tasks has been added to do it.

The watch task has also been updated meaning that
`./node_modules/.bin/gulp watch` correctly rebuilds all assets on
change.